requirements.txt
openai
openai-whisper
streamlit
moviepy
setuptools-rust
